Updated openstack/openstack Project: openstack/neutron  6a0d52510add23a51eccf9a1b6a4048a74a9bc9d
Allow vlan type usage for OpenDaylight ml2

The initial OpenDaylight integration with Openstack did not support
vlan isolation so it was not included as a valid type. This change
adds the vlan type as allowed.

Also modified tests to include vlan as supported in the
check_segment().

DocImpact:
Use of VLANs with ML2 and the OpenDaylight mechanism driver
requires OpenDaylight Helium or newer to be installed.

Closes-Bug: #1315475
Change-Id: I52e3c9bfdc93b8786c58954beca105e7498e3f40