Updated openstack/openstack Project: openstack/python-swiftclient  916f5cbd61eac86527b7a9b867ea1ddb3c84cdf5
Relax requirement for tenant_name in get_auth()

get_auth() in client.py raises an exception if
tenant_name is not included in the os_options
dict. This is overly constrained since tenant_id
is equally sufficient. This patch modifies
get_auth to require either tenant_name
or tenant_id.

Change-Id: Ibbcda1704637eb887efa5895579d260a1e072327