Updated openstack/openstack Project: openstack-infra/devstack-gate  607ca500e6df65df8d8126938df88d43308fc61b
enable heat in grenade

this enables heat in the grenade jobs. Because we need to explicitly
tell tempest what services to test, this doesn't actually test
heat (that will depend on additional grenade changes).

This is therefore a largely no-op change in normal runs which will
make it possible for the heat team to add real grenade support.

Change-Id: I6cfb2e4abf43abf91f96e85e82fcdd1fee3f9470