Updated openstack/openstack Project: openstack/horizon  8314fb13670cf32f3c8f5076fe222f2f0f3a6692
not display loading image for rescued instances

If, any of the instances are rescued or shelved through CLI
those instances status are displayed as rescue or shelved offloaded
with a loading image.

Displaying the loading image for this two status are not correct.

Displaying the loading image is removed for
the rescue and shelved offloaded status.

Change-Id: Ic29da8bc4a6bc75ae9269280f5b28e9ad6a63c2e
Closes-Bug: #1318558
Co-Authored-By: Mizielski Robert <robert.mizielski@cloudwatt.com>