Updated openstack/openstack Project: openstack/nova  2c678b3afbc2bb3913063c0ebfaf1b0ea6653c57
remove unneeded call to network_api on detach_interface

When detaching an interface from an instance we call the network_api
which results in calling neutron and updating the info cache. We do not
actually need to do this as we can get the nw_info directly from the
instance's info cache which removes the call to neutron.

Closes-bug: 1294886

Change-Id: Ifbfff958146a4a01b2d9c656180ef4a95e154312