Updated openstack/openstack Project: openstack/requirements  516ef6dee40e7e86bb43e4608f529845454a9309
Add pyca/cryptography to global requirements

barbican would like to convert some of its internal encryption
to using the cryptography project from PyCA. Specifically our
dev plugin would benefit from a switch to cryptography as it
allows us to use AEAD modes. We will also need this in the
future as cryptography adds X509 and CMS support.

The project supports Python 2.6-3.4 as well as PyPy and is being
packaged for FreeBSD, Debian, and Ubuntu.

Additionally, this is already a second-level requirement when
using PyOpenSSL.

Change-Id: Ic0d508faef9de347bf7cb2436a433096779cf8ac