Updated openstack/openstack Project: openstack/django_openstack_auth  bf9c7f37d238d5d82205147cf4bed1995375d620
Redirect the user if they're already logged in

Similar to when the domain root url is accessed, if the user is
already logged in do not show them the login page nor ask for their
credentials. This avoids users being misled into thinking they can
open multiple sessions in parallel, and is in line with how most web
applications handle this.

Change-Id: Ibd37b9c488d65cf54b156f23db4fa04f019d8092
Closes-Bug: #1308637