Updated openstack/openstack Project: openstack/nova  41f3003c7bf175dd0734fbe5292cb0a0da6f2f49
XenAPI: Add logging information for cache/download duration

We would like more explicit information on when images are cached.
This commit logs information on whether we are attempting to cache,
whether we downloaded, and how long it took to download.  This
log information can then be used for further processing elsewhere.

Change-Id: Ia971008fabec06c4ba8fc1db4b67474f52604b17