Updated openstack/openstack Project: openstack/nova  1fb9e99c187c8dfa984e0a2362ade0e31b971a4a
Add new ec2 instance db API calls

The existing API calls for the EC2 instance ID mappings only returned
the EC2 ID or the UUID, not the full mapping.  We really want APIs
that return the whole mapping for use with a nova object.  This patch
adds those APIs which will be used by the InstanceMapping nova object.

Change-Id: Ie1e5aa1c80c92cc65c6a1e062922f0375b742ea6