Updated openstack/openstack Project: openstack/tempest  493996ed087e3a72f7c5257b20535c42aa0adb12
Fix security_group_create_get_delete testcase

As per the description of test_security_group_create_get_delete,
the test case is supposed to check the security group creation,
followed by retrieval and then deletion of security group.
Deletion of security group is missing in the testcase.

This patch also addresses the issue in the test case,
test_security_groups_create_list_delete operating on
self.security_groups which includes all the security_groups
created by previous test-cases. self.security_groups is not
updated when a security_group is deleted, as we use
self.client.delete_security_group for deletion. OTOH relying
on self.security_groups is wrong because the objective of test
case is to validate only the security_groups that are created in
the current testcase and not the global list of Security_groups.

Change-Id: I3221d41003b577f0a49a7515f1c90e8328cb0dd1
Closes-Bug: #1318644