Updated openstack/openstack Project: openstack/nova  1f77df492f0b971a30656a5cbb5abedba0007cea
Call _validate_instance_group_policy in _build_and_run_instance

The _validate_instance_group_policy check was added to the boot process
in run_instance but was not added to build_and_run_instance which will
be replacing run_instance.  This ports it over to retain parity between
the two code paths.

Part of bp remove-cast-to-schedule-run-instance

Change-Id: I197794ffe04d7460cf67216b9ae67ed7c5f790d9