Updated openstack/openstack Project: openstack/python-heatclient  966470d82e480ec96ce27db8eb06d4f143e158cd
Add CLI commands to interact with software-config

This adds the following heat cli commands for interacting
with the config and deployment REST APIs outside of a heat
template.

It adds the following commands:
- config-create
- config-show
- config-delete
- deployment-show
- deployment-metadata
- deployment-delete

This change does not include deployment-create or deployment-update
commands. These will come in a later change once it becomes obvious
how to discover the stack_user_project_id to pass.

Partial-Bug: #1319986
Change-Id: I405560f7bc085cc35d791cd1218691e7999f9502