Updated openstack/openstack Project: openstack/trove  42e11d8c0021c9d9529e60d99c8f3e3974447e6a
Consider datastore version when generating configs

This commit upgrades the bit of code which determines the datastore
config to render. Currently that code only acknowledges the datastore
manager, but with this config it's possible (but not required) to let
it use a unique config based on version or even flavor.

Author: Tim Simpson <tim.simpson@rackspace.com>
Co-Authored-By: Craig Vyvial <cp16net@gmail.com>

Change-Id: I17ac1ab03bad4ad3606e0cf3d945ca8d7313f96c
Implements: blueprint datastore-configs-cant-use-multiple-versions