Updated openstack/openstack Project: openstack/cinder  2f941d9e04b6403f393902c184d3e7eb9bb7a8a1
Remove create_from* functor jump table

To make it easier to debug during runtime which method was
called remove the dictionary type -> functor mapping and
use a set of if statements instead. This makes it so that
if one of these methods fails we can gather more information
from the traceback than what was previously possible.

Change-Id: Iee2b0b90fe57734966d415c053bbc967335d5509