Updated openstack/openstack Project: openstack-infra/devstack-gate  5cab8f405772d4658eb144ae041458bfa78e0a5c
localize variables

there was coupling between tests because of variables that might get
set in one test and still exist in following ones (but not be declared
there). By localizing all our variables we can isolate ourselves.

To prove that we did this correctly, scramble the test order at the
end (more specifically alphabetize it), so that it has no relation
to the declare order.

Change-Id: I6a52b85b909e54d692abe6d2cc2223dc2e9154e4