Updated openstack/openstack Project: openstack/tempest  aa7eb963cb523275984295b6a073c667c5606903
Update version string to follow new convention

Since adopting branchless we've moved to a different "release" model
were we just tag the tree at set points. We're using an incrementing
integer to name these tags. Since the first tag, 1 was pushed to
correspond with the start of branchless tempest. The current working
tag is 2.

Change-Id: I93be84da9bf2a09e9ca9c2f4231f1a4a75da3af9