Updated openstack/openstack Project: openstack/python-keystoneclient  82cf5d22bd2df967913d309243ee877b699f5f37
Auth Plugin invalidation

To allow session to re-fetch a token on an Unauthorized call we add an
invalidate method to auth plugins that is expected to flush all the
current authentication data from the plugin such that it will be
refreshed on next request.

This is then used to reissue requests from session when an Unauthorized
is called.

Change-Id: I98fa76fd67e97dc0a8c1ec0bf734792c337b5177
blueprint: keystoneclient-auth-token