Updated openstack/openstack Project: openstack/neutron  9a7a5c7e4c5f818d64180ed5d6514dba3f4dbdaa
Fixes bugs for requests sent to SDN-VE controller

This patchset fixes a couple of bugs wrt processing requests
before sending them to the backend controller and adjusts the
requests based on the controller requirements. It also corrects
typos, add quotas and changes the default for a configuration
parameter.

Change-Id: I4b64c2b49ff4854949afc1e54cba1057f376b058
Closes-Bug: #1311260
DocImpact: The default tenant type is chganged from OF to OVERLAY