Updated openstack/openstack Project: openstack/nova  308996e00a50201547c0ac74d4f1e1710736d472
Inject expected results for IBM Power when testing bus

There's no support for ide in IBM Power, and this test is
failing when it's being running in Power.
This patch inject the expected value in
test_xml_disk_bus_ide_and_virtio and
test_xml_disk_bus_ide when it's tested in Power
platform, so, when architecture is Power, the injected value
replace the ide by scsi.

Change-Id: I2db1ade0e049c3471989b121b9497bc5124ea567