Updated openstack/openstack Project: openstack/horizon  78da016d475219ac30d1b98ffd4c066d81355764
Make selenium tests pass in WebKit

The inline editing selenium tests aren't handling the behaviour of WebKit
where it makes the text on a focused input field selected. Instead of
prepending some text to the field, the input keys completely replace it.

Fix this by sending the Home key before sending the text.

Closes-bug: #1321514
Change-Id: I003bf50eff3eb8ebb53125bfab03caf607e92291