Updated openstack/openstack Project: openstack-dev/devstack  a0ced4df0073e8c2405f41dbb8b50dbfb245eff8
chown stack/data/nova/networks to $STACK_USER

$USER may be undefined (resulting in an error exit from stack.sh)
or may be the current $USER rather than the target $STACK_USER
(which may or may not be the same) which could lead to errors
later.

This change brings about consistency: the rest of stack.sh always
uses $STACK_USER.

This same problem may be present in lib/marconi.

Change-Id: Idff2bee88de4e1da11cd5765a03646caf3ede24f