Updated openstack/openstack Project: openstack/horizon  07c097f1ff5147eeff5f498482adc5159ae2fe74
Bug fix on neutron's API to return the correct target ID

In the class FloatingIpManager, the function filters the
resulting list of ports with all available floating ips.
The function now returns the target ID of a port with a
floating IP associated to it. Added a new function to get
this to pass. In the next phase is to refactor the
floating ip api to allow a finer granularity. This will
be useful later for managing floating ips.

Conflicts:
	openstack_dashboard/dashboards/project/instances/tables.py

Change-Id: I02286dbbb60318b0157992622ffdeeae45a71bc1
Closes-Bug: 1265872
(cherry picked from commit 86f87944900e6e713d2baaacec18f55cbe81d878)