Updated openstack/openstack Project: openstack/nova  73d95d2817452f78c468ce162265be431f625c8a
Sync periodic_task from oslo-incubator

We have not updated periodic_task in over six months and we should have
commit 47c9d60657be4ad58afa3c7e3ef88885a595c4c3 in nova.

Note that this does not include the gettextutils and log dependencies
since there are not functional changes in those modules needed for the
periodic_task changes synced in *and* more importantly, the changes
to gettextutils and log require pervasive changes to nova which should
happen when nova integrates with the oslo-i18n library for
blueprint i18n--messages.

Further note that this does not include jsonutils due to some
issues introduced with a change for python 2.6 that impacts some of
the unit tests in Nova and how strings are encoded with simplejson.
The details for that issue are in bug 1314129.  The jsonutils changes
are not related to the periodic_task changes being synced in so the
dependency is not functionally required.

periodic_task:
c63fd5a Make unspecified periodic spaced tasks run on default interval
f0dd798 Remove rendundant parentheses of cfg help strings
fcf517d Update oslo log messages with translation domains
051b9f3 Refactor unnecessary arithmetic ops in periodic_task
674cdaf Refactor if logic in periodic_task
b6b82c5 Use timestamp in periodic tasks
47c9d60 Don't share periodic_task instance data in a class attr
8b2b0b7 Use hacking import_exceptions for gettextutils._
c5a1088 Typos fix in db and periodic_task module
12bcdb7 Remove vim header

Related-Bug: #1319232

Change-Id: I5534e85ae8791c7c464357f7fd57b986ff74d092