Updated openstack/openstack Project: openstack/nova  bba9be040ff617eca9129e0169583284329ac77c
Treat instance objects as objects in rescue API flow

The compute API and manager are getting instance objects now but
treating them like primitive dicts still in several places, so this
cleans that up.

Also adds the instance to the warning log message in _get_rescue_image
since (a) I'm already in there and (b) it adds context for the message,
which is useful.

Cleans up the related tests so they are using objects as well to verify
the changes to compute api/manager are correct.

Part of blueprint compute-manager-objects-juno

Change-Id: Ibe4473fc08fd68676aaa1e1c9ba69161fc5f527e