Updated openstack/openstack Project: openstack/nova  d183110d3d80e2bcabebbb68188f1deab8be3afe
Treat instance objects as objects in unrescue API flow

The compute API and manager are getting instance objects now but
treating them like primitive dicts still in several places, so this
cleans that up.

Cleans up the related tests so they are using objects as well to verify
the changes to compute api/manager are correct.

Part of blueprint compute-manager-objects-juno

Change-Id: I81fa858a821a2137fce5970905b9b1c71c95304b