Updated openstack/openstack Project: openstack/horizon  2f57805e2d0fd1439b429a53d1c7ed890400b691
wrap_list is not honored for not-editable cells

wrap_list attribute is used to add an enclosing <ul> tag to table cell
content.
_data_table_cell.html lacks the code to add that element if the cell content
is not editable (i.e. inline_edit_available is False or update_allowed is
False).

This patch add that missing code and some tests to ensure wrap_list works as
expected.

Change-Id: I261c345790b4dd2c8c6865d98e6ddad136ae1a72
Closes-Bug: #1316666