Updated openstack/openstack Project: openstack/tempest  e11488ac9291519bbc7ae0213224ba50a3295161
allow 'main' as valid name for find_test_caller

find_test_caller assumed that the rest clients were only ever run
under testr. With the introduction of javelin into the tempest
tree this will not be true. 'main' is a completely valid name
for well formed programs, so make that an option as well.

Prereq for blueprint:javelin2

Change-Id: I9a591eaa1cf4dabba58f06a64814611a05a51365